Moved to single server architecture (TODO: Implement server-side routing, fix AWS double click error)
