+++ package.json
@@ -44,8 +44,8 @@
     "eject": "react-scripts eject",
     "populateDB": "node server/database/PopulateDB.js",
     "postinstall": "husky install",
-    "server": "node server/Server.js",
-    "start": "yarn lint; react-scripts start",
+    "start": "yarn lint; node server/Server.js",
+    "startLocal": "yarn lint; react-scripts start",
     "test": "yarn lint; react-scripts test",
     "lint": "eslint --fix ."
   },
