+++ src/common/Managers/EndpointManager.js
@@ -1,24 +1,30 @@
 const axios = require('axios');
-const root = `http://localhost:3001`; // make sure this matches up with backend
+const root = '/api'; // make sure this matches up with backend
 
 /**
  * A file for storing endpoint calls. Let the caller handle errors, so it can display appropriate errors to client
  */
 
+/*
+ * Auth endpoints
+ */
+const authRoot = `${root}/auth`;
+
 export function validateGoogleUser(id_token) {
-  return axios.post(`${root}/gauth`,{id_token});
+  return axios.post(`${authRoot}/gauth`,{id_token});
 }
 
 /*
-* Main endpoints
+* Job endpoints
 */
+const jobRoot = `${root}/jobs`;
 
 export function submitJob(id_token, jobName, maxJobTime, targetColumn, targetColumnName, dataset) {
-  return axios.post(`${root}/submitJob`, {id_token, jobName, maxJobTime, targetColumn, targetColumnName, dataset});
+  return axios.post(`${jobRoot}/submitJob`, {id_token, jobName, maxJobTime, targetColumn, targetColumnName, dataset});
 }
 
 export function getJobs(id_token) {
-  return axios.post(`${root}/jobs`, {id_token});
+  return axios.post(`${jobRoot}`, {id_token});
 
 }
 
@@ -28,7 +34,7 @@ export function deleteJob(id_token, jobId) {
 
   return axios({
     method: "delete",
-    url: `${root}/deleteJob`,
+    url: `${jobRoot}/deleteJob`,
     data: d
   });
 }
@@ -37,24 +43,25 @@ export function deleteJob(id_token, jobId) {
 * AWS endpoints
 */
 
+const awsRoot = `${root}/aws`;
+
 export function registerAWS(region, accessKeyId, secretAccessKey) {
   return axios({
     method: "post",
-    url: `${root}/registerAWS`,
+    url: `${awsRoot}/register`,
     data: {region, credentials: {accessKeyId, secretAccessKey}}
   });
 }
 
 export function listBuckets() {
-  return axios.get(`${root}/listBuckets`)
+  return axios.get(`${awsRoot}/listBuckets`)
     .then(res => res.data);
 }
 
 export function listObjects(bucketName) {
-  console.log(`Listing objects for: ${bucketName}`);
   return axios({
     method: "post",
-    url: `${root}/listObjects`,
+    url: `${awsRoot}/listObjects`,
     data: {bucketName: bucketName}
   })
     .then(res => res.data);
@@ -64,7 +71,7 @@ export function getObject(bucketName, key) {
   console.log(`Retrieving file ${key}`);
   return axios({
     method: "post",
-    url: `${root}/getObject`,
+    url: `${awsRoot}/getObject`,
     data: {bucketName, key}
   })
     .then(res => res.data);
